Move InMemoryExceptionChunk to separate file
