+++ src/include/storage/store/column_chunk_data.h
@@ -12,6 +12,7 @@
 #include "storage/enums/residency_state.h"
 #include "storage/store/column_chunk_metadata.h"
 #include "storage/store/column_reader_writer.h"
+#include "storage/store/in_memory_exception_chunk.h"
 
 namespace kuzu {
 namespace evaluator {
